BREAK: All blocks that have been placed by a player will not count in the quest progression.
When a player place a block, this block is getting a FixedMetadataValue, with the player name.
If the player break a block that have a FixedMetadataValue with his name, the quest will not progress.

PLACE:
All blocks that have been broken by a player will not count in the quest progression.

When a player break a block, the drops of this block are getting a PersistentDataContainer, with the player name.
If the player place a block from an item that have a PersistentDataContainer with his name, the quest will not progress.

PICKUP:
All items that have been dropped by a player will not count in the quest progression.

Every item dropped by a player are getting a PersistentDataContainer, with the player name.
If the player pickup an item that have a PersistentDataContainer with his name, the quest will not progress.