(11/08) Added support for all plugin placeholders in quest descriptions (%drawIn%, %achieved%...). The use_placeholders: true option is required. (11/08) Added use_custom_furnace_results option, to be enabled if the furnace can give items with a particular custom model data. (11/08) Added support for all custom recipes (and custom items as results) in CRAFT quests. (11/08) Added the ability to specify several entity names for CUSTOM_MOBS quests. (11/08) Error added for incorrect potion type in quest configuration.
(14/08) Added support for items from MMOItems for quests and the interface, in the same form as previous integrations (e.g. required_item: "mmoitems:<category>:<id>").

(17/08) Added the disable_logs option in the configuration file.
(17/08) Added the %status% placeholder, that you can use in the quest description. It will return the achieved message if the quest is completed, or the progress message if not.
(17/08) Added the disable_status option in player_interface, to disable the automatically displayed messages below the quest.

(17/08) Fixed an error when placing blocks while the antiglitch is enabled.
(17/08) Fixed a data backup problem when the server is shut down.

(18/08) Fixed an error related to the last snapshot that caused the plugin to crash on each reload.